+++ HISTORY.md
@@ -4,6 +4,7 @@
   * Accept `RegExp` object for `hostname`
   * Remove old invocation of `server.onvhost`
   * String `hostname` with `*` behave more like SSL certificates
+  * Support IPv6 literal in `Host` header
 
 1.0.0 / 2014-03-05
 ==================